Move image download to outside
