+++ src/marqo/core/inference/image_download.py
