+++ src/marqo/core/inference/models/abstract_clip_model.py
@@ -4,8 +4,8 @@ from PIL import UnidentifiedImageError
 
 from marqo.core.inference.models.abstract_embedding_model import AbstractEmbeddingModel
 from marqo.s2_inference.types import *
-from marqo.core.inference.models.image_download import (_is_image, format_and_load_CLIP_images,
-                                                        format_and_load_CLIP_image)
+from marqo.core.inference.image_download import (_is_image, format_and_load_CLIP_images,
+                                                 format_and_load_CLIP_image)
 from marqo.s2_inference.logger import get_logger
 import torch
